+++ src/components/PlantCard/PlantCard.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import PlantDetails from '../PlantDetails/PlantDetails.jsx';
-import fetchPlantDetails from '../PlantDetails/fetchPlantDetailsApi.jsx';
+import fetchPlantDetails from '../shared/utils/fetchPlantDetailsApi.jsx';
 import { db } from "../../firebase/firebase.js";
 import { collection, addDoc, getDocs, serverTimestamp } from "firebase/firestore";
 import { useAuth } from "../Authentication/AuthContext";
