+++ src/components/shared/utils/fetchPlantDetailsApi.jsx
